refactor serialization
